+++ imports/startup/server/register-api.js
@@ -11,3 +11,6 @@ import '../../api/quotes/quoteMethods';
 import '../../api/quotes/server/quotePublications';
 import '../../api/rates/rateMethods';
 import '../../api/shipments/shipmentMethods';
+import '../../api/stakeholders/stakeholderMethods';
+import '../../api/suppliers/supplierCollection';
+import '../../api/suppliers/supplierMethods';