+++ imports/state/reducers/list/listReducers.js
@@ -1,13 +1,13 @@
 import { set } from 'lodash/fp';
 import * as ACTION_TYPES from '../../actions/actionTypes';
 
-export const defaultCustomerListState = {
+export const defaultListState = {
   filter: '',
 };
 
-export const customerList = (state = defaultCustomerListState, action = { type: '' }) => {
+export const list = (state = defaultListState, action = { type: '' }) => {
   switch (action.type) {
-    case ACTION_TYPES.SET_CUSTOMER_LIST_FILTER:
+    case ACTION_TYPES.SET_LIST_FILTER:
       return set('filter', action.filter, state);
     default:
       return state;