+++ imports/state/reducers/list/listReducers.tests.js
@@ -5,23 +5,21 @@ import { Meteor } from 'meteor/meteor';
 import { chai } from 'meteor/practicalmeteor:chai';
 import deepFreeze from 'deep-freeze';
 
-import { customerList, defaultCustomerListState } from './customerListReducers';
+import { list, defaultListState } from './listReducers';
 import * as ACTION_TYPES from '../../actions/actionTypes';
 
 if (Meteor.isClient) {
-  describe('Customer List Reducer', () => {
+  describe('List Reducer', () => {
     chai.should();
-
     it('defaults each field to blank', () => {
-      customerList().should.eql(defaultCustomerListState);
-      defaultCustomerListState.filter.should.equal('');
+      list().should.eql(defaultListState);
+      defaultListState.filter.should.equal('');
     });
-
-    it('set the filter', () => {
+    it('sets the list filter', () => {
       const stateBefore = { filter: 'a' };
-      const action = { type: ACTION_TYPES.SET_CUSTOMER_LIST_FILTER, filter: 'b' };
+      const action = { type: ACTION_TYPES.SET_LIST_FILTER, filter: 'b' };
       deepFreeze(stateBefore);
-      customerList(stateBefore, action).filter.should.equal('b');
+      list(stateBefore, action).filter.should.equal('b');
     });
   });
 }