+++ imports/ui/editors/EditCustomer.tests.jsx
@@ -18,165 +18,117 @@ if (Meteor.isClient) {
     afterEach(() => {
       wrapper.unmount();
     });
-
     describe('Structure', () => {
       it('renders a component', () => {
         wrapper.exists().should.equal(true);
       });
-
       it('renders a content', () => {
-        wrapper
-          .childAt(0)
-          .hasClass('content')
-          .should
-          .equal(true);
+        wrapper.childAt(0).hasClass('content').should.equal(true);
       });
-
       it('renders a content footer accent', () => {
         wrapper
           .childAt(1)
-          .contains((
-            <div className="content-footer-accent customers-footer-accent" />
-          ))
-          .should
-          .equal(true);
+          .contains(
+            <div className="content-footer-accent customers-footer-accent" />,
+          )
+          .should.equal(true);
       });
-
       it('renders a process header', () => {
         wrapper
           .find('.content')
           .childAt(0)
           .hasClass('process-header')
-          .should
-          .equal(true);
+          .should.equal(true);
       });
-
       it('renders a title in the process header', () => {
         wrapper
           .find('.process-header')
           .childAt(0)
           .contains(<div className="title">NEW CUSTOMER</div>)
-          .should
-          .equal(true);
+          .should.equal(true);
       });
-
       it('renders a link back to the customer list', () => {
-        wrapper
-          .find('.process-header')
-          .childAt(1)
-          .name()
-          .should
-          .equal('Link');
-        wrapper
-          .find('Link')
-          .props()
-          .to
-          .should
-          .equal('/customers');
+        wrapper.find('.process-header').childAt(1).name().should.equal('Link');
+        wrapper.find('Link').props().to.should.equal('/customers');
       });
-
       it('renders a customer list button in the link', () => {
         wrapper
           .find('Link')
-          .contains((
-            <button className="button-primary">BACK TO CUSTOMER LIST</button>
-          ))
-          .should
-          .equal(true);
+          .contains(
+            <button className="button-primary">BACK TO CUSTOMER LIST</button>,
+          )
+          .should.equal(true);
       });
-
       it('renders a panel', () => {
         wrapper
           .find('.content')
           .childAt(1)
           .hasClass('panel')
-          .should
-          .equal(true);
+          .should.equal(true);
       });
-
       it('renders a name input field', () => {
         wrapper
           .find('.panel .vertical-input-group')
           .at(1)
           .childAt(0)
           .contains(<span className="label">NAME</span>)
-          .should
-          .equal(true);
+          .should.equal(true);
         wrapper
           .find('.panel .vertical-input-group')
           .at(1)
           .childAt(1)
           .name()
-          .should
-          .equal('input');
+          .should.equal('input');
       });
-
       it('renders an address textarea field', () => {
         wrapper
           .find('.panel .vertical-input-group')
           .at(2)
           .childAt(0)
           .contains(<span className="label">ADDRESS</span>)
-          .should
-          .equal(true);
+          .should.equal(true);
         wrapper
           .find('.panel .vertical-input-group')
           .at(2)
           .childAt(1)
           .name()
-          .should
-          .equal('textarea');
+          .should.equal('textarea');
       });
-
       it('renders a currency input field', () => {
-        wrapper
-          .find('.panel .vertical-input-group')
-          .at(4)
-          .childAt(0)
-          .contains(<span className="label">CURRENCY</span>)
-          .should
-          .equal(true);
         wrapper
           .find('.panel .vertical-input-group')
           .at(4)
           .childAt(1)
           .name()
-          .should
-          .equal('input');
+          .should.equal('CurrencyField');
       });
-
       it('renders an email address input field', () => {
         wrapper
           .find('.panel .vertical-input-group')
           .at(3)
           .childAt(0)
           .contains(<span className="label">EMAIL ADDRESS</span>)
-          .should
-          .equal(true);
+          .should.equal(true);
         wrapper
           .find('.panel .vertical-input-group')
           .at(3)
           .childAt(1)
           .name()
-          .should
-          .equal('input');
+          .should.equal('input');
       });
-
       it('renders a branch input field', () => {
         wrapper
           .find('.panel .vertical-input-group')
           .at(0)
           .childAt(0)
           .contains(<span className="label">BRANCH</span>)
-          .should
-          .equal(true);
+          .should.equal(true);
         wrapper
           .find('.panel .vertical-input-group')
           .at(0)
           .childAt(1)
           .name()
-          .should
-          .equal('BranchField');
+          .should.equal('BranchField');
       });
     });
   });